VideoBufferDRMPRIME: extract class
